### 2018-10-19 #1032 @terrytangyuan
@arfon Could you be the editor?
