### 2018-10-19 #1036 @whedon
[REVIEW]: dml: Distance Metric Learning in R
**Submitting author:** @terrytangyuan (<a href="http://orcid.org/0000-0001-5243-233X">Yuan Tang</a>)
**Repository:** <a href="https://github.com/terrytangyuan/dml" target ="_blank">https://github.com/terrytangyuan/dml</a>
**Version:** v1.1.0
