### 2018-10-19 b67bace
Additional references
