+++ paper/paper.bib
@@ -57,3 +57,42 @@
   year = {2018},
   url = {https://CRAN.R-project.org/package=autoplotly},
 }
+
+@InProceedings{rcapaper,
+  title = {Adjustment learning and relevant component analysis},
+  author = {Shental, Noam and Hertz, Tomer and Weinshall, Daphna and Pavel, Misha},
+  booktitle = {European Conference on Computer Vision},
+  pages = {776--790},
+  year = {2002},
+  organization = {Springer}
+}
+
+@InProceedings{dcapaper,
+  title={Fast semi-supervised discriminative component analysis},
+  author={Peltonen, Jaakko and Goldberger, Jacob and Kaski, Samuel},
+  booktitle={Machine Learning for Signal Processing, 2007 IEEE Workshop on},
+  pages={312--317},
+  year={2007},
+  organization={IEEE}
+}
+
+@Article{lfdaoriginalpaper,
+  title={Dimensionality reduction of multimodal labeled data by local fisher discriminant analysis},
+  author={Sugiyama, Masashi},
+  journal={Journal of machine learning research},
+  volume={8},
+  number={May},
+  pages={1027--1061},
+  year={2007}
+}
+
+@Article{semilfdapaper,
+  title={Semi-supervised local Fisher discriminant analysis for dimensionality reduction},
+  author={Sugiyama, Masashi and Id{\'e}, Tsuyoshi and Nakajima, Shinichi and Sese, Jun},
+  journal={Machine learning},
+  volume={78},
+  number={1-2},
+  pages={35},
+  year={2010},
+  publisher={Springer}
+}
