+++ paper/paper.md
@@ -8,7 +8,7 @@ authors:
   name: Tao Gao
 - affiliation: 3
   name: Nan Xiao
-date: "17 October 2018"
+date: "19 October 2018"
 output:
   pdf_document: default
   html_document:
