+++ paper/paper.bib
@@ -6,6 +6,24 @@
     url = {http://CRAN.R-project.org/package=ggfortify}
 }
 
+@InProceedings{dcapaper,
+  title={Fast semi-supervised discriminative component analysis},
+  author={Peltonen, Jaakko and Goldberger, Jacob and Kaski, Samuel},
+  booktitle={Machine Learning for Signal Processing, 2007 IEEE Workshop on},
+  pages={312--317},
+  year={2007},
+  organization={IEEE}
+}
+
+@InProceedings{rcapaper,
+  title = {Adjustment learning and relevant component analysis},
+  author = {Shental, Noam and Hertz, Tomer and Weinshall, Daphna and Pavel, Misha},
+  booktitle = {European Conference on Computer Vision},
+  pages = {776--790},
+  year = {2002},
+  organization = {Springer}
+}
+
 @Article{rjggfortify,
   author = {Tang, Yuan and Horikoshi, Masaaki and Li, Wenxuan},
   title = {ggfortify: Unified Interface to Visualize Statistical Result of Popular R Packages},
@@ -19,14 +37,14 @@
 
 @Manual{dmlpkg,
   title = {dml: Distance Metric Learning in R},
-  author = {Yuan Tang and Tao Gao and Nan Xiao},
+  author = {Tang, Yuan and Gao, Tao and Xiao, Nan},
   year = {2015},
   note = {R package version 1.1.0},
   url = {https://CRAN.R-project.org/package=dml},
 }
 
 @Article{lfdapaper,
-  author = {{Tang} and {Yuan} and {Li} and {Wenxuan}},
+  author = {Tang, Yuan and Li, Wenxuan},
   title = {lfda: An R Package for Local Fisher Discriminant Analysis and Visualization},
   journal = {arXiv preprint arXiv:1612.09219},
   year = {2016},
@@ -34,7 +52,7 @@
 }
 
 @Manual{lfdapkg,
-  author = {Yuan Tang},
+  author = {Tang, Yuan},
   title = {Local Fisher Discriminant Analysis},
   year = {2017},
   url = {https://CRAN.R-project.org/package=lfda},
@@ -52,29 +70,12 @@
 }
 
 @Manual{autoplotlypkg,
-  author = {Yuan Tang},
+  author = {Tang, Yuan},
   title = {autoplotly: An R Package for Automatic Generation of Interactive Visualizations for Statistical Results},
   year = {2018},
   url = {https://CRAN.R-project.org/package=autoplotly},
 }
 
-@InProceedings{rcapaper,
-  title = {Adjustment learning and relevant component analysis},
-  author = {Shental, Noam and Hertz, Tomer and Weinshall, Daphna and Pavel, Misha},
-  booktitle = {European Conference on Computer Vision},
-  pages = {776--790},
-  year = {2002},
-  organization = {Springer}
-}
-
-@InProceedings{dcapaper,
-  title={Fast semi-supervised discriminative component analysis},
-  author={Peltonen, Jaakko and Goldberger, Jacob and Kaski, Samuel},
-  booktitle={Machine Learning for Signal Processing, 2007 IEEE Workshop on},
-  pages={312--317},
-  year={2007},
-  organization={IEEE}
-}
 
 @Article{lfdaoriginalpaper,
   title={Dimensionality reduction of multimodal labeled data by local fisher discriminant analysis},
