### 2018-10-19 #1036 @terrytangyuan
Thanks. @arfon It seems like the "`R`"s got lower-cased in the references, which didn't happen before for my previous publication through JOSS. Do you know if anything has changed?
