### 2018-10-19 #1036 @arfonsmith
I know we've changed citation style a few months ago. I'm away from my
computer until Sunday evening but will try and take a look then.

On Fri, Oct 19, 2018 at 10:52, Yuan (Terry) Tang <notifications@github.com>
wrote:

> Thanks. @arfon <https://github.com/arfon> It seems like the "R"s got
> lower-cased in the references, which didn't happen before for my previous
> publication through JOSS. Do you know if anything has changed?
>
> —
>
>
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/openjournals/joss-reviews/issues/1036#issuecomment-431389953>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AGjvLgswlaONp12p8DUfVJchM5Lj0jDkks5umec3gaJpZM4XwEU2>
> .
>
