### 2018-10-19 terrytangyuan/dml#27 @joethorley
please clarify what is meant by state-of-the art
https://github.com/openjournals/joss-reviews/issues/1036
