### 2018-10-19 terrytangyuan/dml#28 @joethorley
please add some examples with R code to replicate to the README
https://github.com/openjournals/joss-reviews/issues/1036
