### 2018-10-20 #1036 @joethorley
@terrytangyuan and @arfon I find the fact that the organizational affiliation are just the names of the organizations to be disconcerting. Is this standard.
