### 2018-10-20 terrytangyuan/dml#32 @joethorley
package scatterplot3d should be in suggests
as its used in examples

https://github.com/openjournals/joss-reviews/issues/1036
