### 2018-10-20 terrytangyuan/dml#33 @joethorley
the date in the DESCRIPTION file needs updating from 2015-08-26
https://github.com/openjournals/joss-reviews/issues/1036
