### 2018-10-20 98ff849
update fields in DESCRIPTION
