+++ DESCRIPTION
@@ -1,13 +1,23 @@
 Package: dml
-Version: 1.1.0.9000
-Date: 2015-08-26
+Type: Package
 Title: Distance Metric Learning in R
-Description: The state-of-the-art algorithms for distance metric learning,
+Version: 1.1.0.9001
+Authors@R: c(
+    person("Yuan", "Tang", email = "terrytangyuan@gmail.com",
+           role = c("aut", "cre"), comment = c(ORCID = "0000-0001-5243-233X")),
+    person("Tao", "Gao", email = "joegaotao@gmail.com", role = c("aut")),
+    person("Nan", "Xiao", email = "me@nanx.me",
+           role = c("aut"), comment = c(ORCID = "0000-0002-0250-5673")))
+Maintainer: Yuan Tang <terrytangyuan@gmail.com>
+Description: State-of-the-art algorithms for distance metric learning,
     including global and local methods such as Relevant Component Analysis,
     Discriminative Component Analysis, Local Fisher Discriminant Analysis, etc.
-    These distance metric learning methods are widely applied in feature extraction,
-    dimensionality reduction, clustering, classification, information retrieval, and
-    computer vision problems.
+    These distance metric learning methods are widely applied in feature
+    extraction, dimensionality reduction, clustering, classification,
+    information retrieval, and computer vision problems.
+License: MIT + file LICENSE
+URL: https://github.com/terrytangyuan/dml
+Encoding: UTF-8
 Depends:
     MASS
 Imports:
@@ -15,10 +25,5 @@ Imports:
 Suggests:
     lintr,
     testthat
-Author: Yuan Tang <terrytangyuan@gmail.com>, Tao Gao <joegaotao@gmail.com>,
-    Nan Xiao <me@nanx.me>
-Maintainer: Yuan Tang <terrytangyuan@gmail.com>
-License: MIT + file LICENSE
-URL: https://github.com/terrytangyuan/dml
 BugReports: https://github.com/terrytangyuan/dml/issues
-RoxygenNote: 5.0.1
+RoxygenNote: 6.1.0
