+++ README.md
@@ -7,20 +7,25 @@
 
 R package for state-of-the-art algorithms for *Distance Metric Learning*, including global and local methods such as *Relevant Component Analysis*, *Discriminative Component Analysis*, *Local Fisher Discriminant Analysis*, etc. These distance metric learning methods are widely applied in feature extraction, dimensionality reduction, clustering, classification, information retrieval, and computer vision problems.
 
-## Install the current release from CRAN:
-```{R}
-install.packages('dml')
+## Installation
+
+Install the current release from CRAN:
+
+```r
+install.packages("dml")
 ```
 
-## Install the latest development version from github:
-```{R}
-devtools::install_github('terrytangyuan/dml')
+Or, try the latest development version from GitHub:
+
+```r
+devtools::install_github("terrytangyuan/dml")
 ```
 
-## Examples:
+## Examples
+
 For examples of Local Fisher Discriminant Analysis, please take a look at the separate package [here](https://github.com/terrytangyuan/lfda). For examples of all other implemented algorithms, please take a look at the dml [package reference manual](https://cran.r-project.org/web/packages/dml/dml.pdf). 
 
-## Brief Intro
+## Brief Introduction
 
 Distance metric is widely used in the machine learning literature. We used to choose a distance metric according to a priori (Euclidean Distance , L1 Distance, etc.) or according to the result of cross validation within small class of functions (e.g. choosing order of polynomial for a kernel). Actually, with priori knowledge of the data, we could learn a more suitable distance metric with (semi-)supervised distance metric learning techniques. dml is such an R package aims to implement the state-of-the-art algorithms for (semi-)supervised distance metric learning. These distance metric learning methods are widely applied in feature extraction, dimensionality reduction, clustering, classification, information retrieval, and computer vision problems.
 
