### 2018-10-20 61cec3c
reformat code with tidyverse style
