+++ R/gdmf.r
@@ -9,9 +9,9 @@
 #'             Each data point is a row in the matrix.
 #' @param simi \code{n * 2} matrix describing the similar constrains.
 #'              Each row of matrix is serial number of a similar pair in the original data.
-#'				For example, pair(1, 3) represents the first observation is similar the 3th observation in the original data.
+#' 				For example, pair(1, 3) represents the first observation is similar the 3th observation in the original data.
 #' @param dism \code{n * 2} matrix describing the dissimilar constrains as \code{simi}.
-#'				Each row of matrix is serial number of a dissimilar pair in the original data.
+#' 				Each row of matrix is serial number of a dissimilar pair in the original data.
 #' @param maxiter numeric, the number of iteration.
 #'
 #' @return list of the GdmDiag results:
@@ -32,7 +32,7 @@
 #'
 #' @note Be sure to check whether the dimension of original data and constrains' format are valid for the function.
 #'
-#' @author Gao Tao <\url{http://www.gaotao.name}>
+#' @author Tao Gao <\url{http://www.gaotao.name}>
 #'
 #' @references
 #' Steven C.H. Hoi, W. Liu, M.R. Lyu and W.Y. Ma (2003).
@@ -68,198 +68,203 @@
 #' color <- gl(2, k, labels = c("red", "blue"))
 #' par(mfrow = c(2, 1), mar = rep(0, 4) + 0.1)
 #' scatterplot3d(data, color = color, cex.symbols = 0.6,
-#'			  xlim = range(data[, 1], newData[, 1]),
-#'			  ylim = range(data[, 2], newData[, 2]),
-#'			  zlim = range(data[, 3], newData[, 3]),
-#'			  main = "Original Data")
+#' 			  xlim = range(data[, 1], newData[, 1]),
+#' 			  ylim = range(data[, 2], newData[, 2]),
+#' 			  zlim = range(data[, 3], newData[, 3]),
+#' 			  main = "Original Data")
 #' # plot GdmFull transformed data
 #' scatterplot3d(newData, color = color, cex.symbols = 0.6,
-#'			  xlim = range(data[, 1], newData[, 1]),
-#'			  ylim = range(data[, 2], newData[, 2]),
-#'			  zlim = range(data[, 3], newData[, 3]),
-#'			  main = "Transformed Data")
+#' 			  xlim = range(data[, 1], newData[, 1]),
+#' 			  ylim = range(data[, 2], newData[, 2]),
+#' 			  zlim = range(data[, 3], newData[, 3]),
+#' 			  main = "Transformed Data")
 #' }
 #'
 GdmFull <- function(data, simi, dism, maxiter = 100) {
-		
-        # Input class checks
-        data <- .checkMatrixInput(functionName = "GdmFull", argName = "data", inputObject = data)
-        simi <- .checkMatrixInput(functionName = "GdmFull", argName = "simi", inputObject = simi)
-        dism <- .checkMatrixInput(functionName = "GdmFull", argName = "dism", inputObject = dism)
-        if (!(is.numeric(maxiter) && length(maxiter) == 1)) {
-            errorTxt <- sprintf("maxiter of GdmFull expects a number. An object of class %s was provided."
-                                , paste(class(maxiter), collapse = "")) 
-            stop(errorTxt)
+
+  # Input class checks
[…]
