+++ R/aaa.R
@@ -1,3 +1,2 @@
-
-#Avoid false positives in R CMD CHECK:
+# Avoid false positives in R CMD CHECK:
 utils::globalVariables(c("Class", "diagonalA"))

[…]
