+++ man/GdmDiag.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gdmd.r
+% Please edit documentation in R/gdmd.R
 \name{GdmDiag}
 \alias{GdmDiag}
 \title{Global Distance Metric Learning}
@@ -12,8 +12,8 @@ GdmDiag(data, simi, dism, C0 = 1, threshold = 0.001)
 Each data point is a row in the matrix.}
 
 \item{simi}{\code{n * 2} matrix describing the similar constrains.
- Each row of matrix is serial number of a similar pair in the original data.
-For example, pair(1, 3) represents the first observation is similar the 3th observation in the original data.}
+Each row of matrix is serial number of a similar pair in the original data.
+   For example, pair(1, 3) represents the first observation is similar the 3th observation in the original data.}
 
 \item{dism}{\code{n * 2} matrix describing the dissimilar constrains as \code{simi}.
 Each row of matrix is serial number of a dissimilar pair in the original data.}
@@ -71,25 +71,25 @@ newData <- result$newData
 color <- gl(2, k, labels = c("red", "blue"))
 par(mfrow = c(2, 1), mar = rep(0, 4) + 0.1)
 scatterplot3d(data, color = color, cex.symbols = 0.6,
-		  xlim = range(data[, 1], newData[, 1]),
-		  ylim = range(data[, 2], newData[, 2]),
-		  zlim = range(data[, 3], newData[, 3]),
-		  main = "Original Data")
+			  xlim = range(data[, 1], newData[, 1]),
+			  ylim = range(data[, 2], newData[, 2]),
+			  zlim = range(data[, 3], newData[, 3]),
+			  main = "Original Data")
 # plot GdmDiag transformed data
 scatterplot3d(newData, color = color, cex.symbols = 0.6,
-		  xlim = range(data[, 1], newData[, 1]),
-		  ylim = range(data[, 2], newData[, 2]),
-		  zlim = range(data[, 3], newData[, 3]),
-		  main = "Transformed Data")
+			  xlim = range(data[, 1], newData[, 1]),
+			  ylim = range(data[, 2], newData[, 2]),
+			  zlim = range(data[, 3], newData[, 3]),
+			  main = "Transformed Data")
 }
 
-}
-\author{
-Gao Tao <\url{http://www.gaotao.name}>
 }
 \references{
 Steven C.H. Hoi, W. Liu, M.R. Lyu and W.Y. Ma (2003).
 Distance metric learning, with application to clustering with side-information.
+}
+\author{
+Tao Gao <\url{http://www.gaotao.name}>
 }
 \keyword{GDM}
 \keyword{distance}
@@ -99,4 +99,3 @@ Distance metric learning, with application to clustering with side-information.
 \keyword{metirc}
 \keyword{metric}
 \keyword{transformation}
-
