+++ man/print.GdmDiag.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gdmd.r
+% Please edit documentation in R/gdmd.R
 \name{print.GdmDiag}
 \alias{print.GdmDiag}
 \title{Print an GdmDiag object}
@@ -14,4 +14,3 @@
 \description{
 Print an GdmDiag object
 }
-
