+++ man/print.gmdf.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gdmf.r
+% Please edit documentation in R/gdmf.R
 \name{print.gmdf}
 \alias{print.gmdf}
 \title{Print a gmdf object}
@@ -14,4 +14,3 @@
 \description{
 Print a gmdf object
 }
-
