+++ man/print.dca.Rd
@@ -14,4 +14,3 @@
 \description{
 Print an dca object
 }
-
