### 2018-10-20 terrytangyuan/dml#33 @road2stat
-good point! The date field is actually not useful anymore so can be safely removed. I've just revamped the DESCRIPTION file in https://github.com/terrytangyuan/dml/commit/98ff849041b25dba3d110acf4a017929b40798c2.
