### 2018-10-20 391c10e
add scatterplot3d as suggests
