+++ R/rca.R
@@ -55,7 +55,6 @@
 #'
 #' @export rca
 #' @importFrom lfda %^%
-#' @import MASS
 #' @importFrom stats cov
 #'
 #' @references
@@ -67,10 +66,10 @@
 #' @examples
 #' \dontrun{
 #' set.seed(1234)
-#' require(MASS)  # generate synthetic Gaussian data
-#' k = 100        # sample size of each class
-#' n = 3          # specify how many class
-#' N = k * n      # total sample number
+#' library("MASS")  # generate synthetic Gaussian data
+#' k = 100          # sample size of each class
+#' n = 3            # specify how many class
+#' N = k * n        # total sample number
 #' x1 = mvrnorm(k, mu = c(-10, 6), matrix(c(10, 4, 4, 10), ncol = 2))
 #' x2 = mvrnorm(k, mu = c(0, 0), matrix(c(10, 4, 4, 10), ncol = 2))
 #' x3 = mvrnorm(k, mu = c(10, -6), matrix(c(10, 4, 4, 10), ncol = 2))
@@ -119,7 +118,7 @@
 #' # The RCA suggested Mahalanobis matrix
 #' rca(x[ , 1:2], chunks)$B
 #' }
-#'
+
 rca <- function(x, chunks) {
   chunkNum <- length(chunks)
   chunkDf <- vector("list", chunkNum)
