+++ R/dca.R
@@ -44,7 +44,6 @@
 #'
 #' @export dca
 #' @importFrom lfda %^%
-#' @import MASS
 #'
 #' @references
 #' Steven C.H. Hoi, W. Liu, M.R. Lyu and W.Y. Ma (2006).
@@ -55,10 +54,10 @@
 #' @examples
 #' \dontrun{
 #' set.seed(123)
-#' require(MASS)  # generate synthetic Gaussian data
-#' k = 100        # sample size of each class
-#' n = 3          # specify how many class
-#' N = k * n      # total sample number
+#' library("MASS")  # generate synthetic Gaussian data
+#' k = 100          # sample size of each class
+#' n = 3            # specify how many class
+#' N = k * n        # total sample number
 #' x1 = mvrnorm(k, mu = c(-10, 6), matrix(c(10, 4, 4, 10), ncol = 2))
 #' x2 = mvrnorm(k, mu = c(0, 0), matrix(c(10, 4, 4, 10), ncol = 2))
 #' x3 = mvrnorm(k, mu = c(10, -6), matrix(c(10, 4, 4, 10), ncol = 2))
