+++ man/GdmFull.Rd
@@ -43,8 +43,8 @@ Be sure to check whether the dimension of original data and constrains' format a
 \examples{
 \dontrun{
 set.seed(123)
-library(MASS)
-library(scatterplot3d)
+library("MASS")
+library("scatterplot3d")
 
 # generate simulated Gaussian data
 k = 100
