+++ NAMESPACE
@@ -8,7 +8,6 @@ export(GdmDiag)
 export(GdmFull)
 export(dca)
 export(rca)
-import(MASS)
 importFrom(lfda,"%^%")
 importFrom(stats,cov)
 importFrom(utils,head)
