### 2018-10-20 5d2c5d4
use canonical form cran url
