+++ R/dml-package.R
@@ -1,0 +1,17 @@
+#' Distance Metric Learning in R
+#'
+#' Solving distance metric learning problems with R.
+#'
+#' \tabular{ll}{
+#' Package: \tab dml\cr
+#' Type: \tab Package\cr
+#' License: \tab MIT}
+#'
+#' @name dml-package
+#' @docType package
+#' @author
+#' Yuan Tang <\email{terrytangyuan@@gmail.com}>
+#' Tao Gao <\email{joegaotao@@gmail.com}>
+#' Nan Xiao <\email{me@@nanx.me}>
+#'
+NULL
