+++ man/dml-package.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/dml-package.R
+\docType{package}
+\name{dml-package}
+\alias{dml-package}
+\title{Distance Metric Learning in R}
+\description{
+Solving distance metric learning problems with R.
+}
+\details{
+\tabular{ll}{
+Package: \tab dml\cr
+Type: \tab Package\cr
+License: \tab MIT}
+}
+\author{
+Yuan Tang <\email{terrytangyuan@gmail.com}>
+Tao Gao <\email{joegaotao@gmail.com}>
+Nan Xiao <\email{me@nanx.me}>
+}
