### 2018-10-20 6811327
fix variable name issue for plotting
