### 2018-10-20 0097c7c
reformat rca example code with tidyverse style
