+++ man/rca.Rd
@@ -59,59 +59,71 @@ Note that any different sets of instances (chunklets),
 }
 \examples{
 \dontrun{
+# generate synthetic Gaussian data
+library("MASS")
+
+k <- 100 # sample size of each class
+n <- 3 # number of classes
+N <- k * n # total number of samples
+
 set.seed(1234)
-library("MASS")  # generate synthetic Gaussian data
-k = 100          # sample size of each class
-n = 3            # specify how many class
-N = k * n        # total sample number
-x1 = mvrnorm(k, mu = c(-10, 6), matrix(c(10, 4, 4, 10), ncol = 2))
-x2 = mvrnorm(k, mu = c(0, 0), matrix(c(10, 4, 4, 10), ncol = 2))
-x3 = mvrnorm(k, mu = c(10, -6), matrix(c(10, 4, 4, 10), ncol = 2))
-x = as.data.frame(rbind(x1, x2, x3))
-x$V3 = gl(n, k)
+x1 <- mvrnorm(k, mu = c(-10, 6), matrix(c(10, 4, 4, 10), ncol = 2))
+x2 <- mvrnorm(k, mu = c(0, 0), matrix(c(10, 4, 4, 10), ncol = 2))
+x3 <- mvrnorm(k, mu = c(10, -6), matrix(c(10, 4, 4, 10), ncol = 2))
+x <- as.data.frame(rbind(x1, x2, x3))
+x$V3 <- gl(n, k)
 
 # The fully labeled data set with 3 classes
-plot(x$V1, x$V2, bg = c("#E41A1C", "#377EB8", "#4DAF4A")[x$V3],
-     pch = c(rep(22, k), rep(21, k), rep(25, k)))
+plot(x$V1, x$V2,
+  bg = c("#E41A1C", "#377EB8", "#4DAF4A")[x$V3],
+  pch = c(rep(22, k), rep(21, k), rep(25, k))
+)
 Sys.sleep(3)
 
-# Same data unlabeled; clearly the classes' structure is less evident
+# The same data unlabeled; clearly the class structure is less evident
 plot(x$V1, x$V2)
 Sys.sleep(3)
 
-chunk1 = sample(1:100, 5)
-chunk2 = sample(setdiff(1:100, chunk1), 5)
-chunk3 = sample(101:200, 5)
-chunk4 = sample(setdiff(101:200, chunk3), 5)
-chunk5 = sample(201:300, 5)
-chks = x[c(chunk1, chunk2, chunk3, chunk4, chunk5), ]
-chunks = list(chunk1, chunk2, chunk3, chunk4, chunk5)
+chunk1 <- sample(1:100, 5)
+chunk2 <- sample(setdiff(1:100, chunk1), 5)
+chunk3 <- sample(101:200, 5)
+chunk4 <- sample(setdiff(101:200, chunk3), 5)
+chunk5 <- sample(201:300, 5)
+chks <- x[c(chunk1, chunk2, chunk3, chunk4, chunk5), ]
+chunks <- list(chunk1, chunk2, chunk3, chunk4, chunk5)
 
 # The chunklets provided to the RCA algorithm
-plot(chks$V1, chks$V2, col = rep(c("#E41A1C", "#377EB8",
-     "#4DAF4A", "#984EA3", "#FF7F00"), each = 5),
-     pch = rep(0:4, each = 5), ylim = c(-15, 15))
+plot(chks$V1, chks$V2,
+  col = rep(c(
+    "#E41A1C", "#377EB8",
+    "#4DAF4A", "#984EA3", "#FF7F00"
+  ), each = 5),
[…]
