### 2018-10-20 b7c5bc3
add contributor guide
