+++ CONTRIBUTING.md
@@ -1,0 +1,23 @@
+# Contributing to dml
+
+:+1::tada: First off, thanks for taking your time to contribute! :tada::+1:
+
+You could contribute to this project by:
+
+1. Filing a bug report or feature request in an issue.
+2. Suggesting a change via a pull request.
+
+## Issues
+
+To file an issue about a possible bug, please try to include:
+
+- Relevant package versions
+- Necessary code and data to reproduce the issue
+
+## Pull Requests
+
+To suggest a change via pull requests, please:
+
+1. Fork the repository into your GitHub account.
+2. Clone the forked repository to local machine, make the changes.
+3. Commit and push the changes to GitHub. Create a pull request.
