### 2018-10-20 53d89de
add links to contributor guide and code of conduct
