### 2018-10-20 72b5c23
fix uppercase issues in bibliography
