+++ paper/paper.bib
@@ -2,8 +2,8 @@
     title = {ggfortify: Data Visualization Tools for Statistical Analysis Results},
     author = {Horikoshi, Masaaki and Tang, Yuan},
     year = {2018},
-    note = {R package version 0.4.4},
-    url = {http://CRAN.R-project.org/package=ggfortify}
+    note = {{R} package version 0.4.4},
+    url = {https://CRAN.R-project.org/package=ggfortify}
 }
 
 @InProceedings{dcapaper,
@@ -26,8 +26,8 @@
 
 @Article{rjggfortify,
   author = {Tang, Yuan and Horikoshi, Masaaki and Li, Wenxuan},
-  title = {ggfortify: Unified Interface to Visualize Statistical Result of Popular R Packages},
-  journal = {The R Journal},
+  title = {ggfortify: Unified Interface to Visualize Statistical Result of Popular {R} Packages},
+  journal = {The {R} Journal},
   year = {2016},
   issue = {2},
   volume = {8},
@@ -39,13 +39,13 @@
   title = {dml: Distance Metric Learning in R},
   author = {Tang, Yuan and Gao, Tao and Xiao, Nan},
   year = {2015},
-  note = {R package version 1.1.0},
+  note = {{R} package version 1.1.0},
   url = {https://CRAN.R-project.org/package=dml},
 }
 
 @Article{lfdapaper,
   author = {Tang, Yuan and Li, Wenxuan},
-  title = {lfda: An R Package for Local Fisher Discriminant Analysis and Visualization},
+  title = {lfda: An {R} Package for Local Fisher Discriminant Analysis and Visualization},
   journal = {arXiv preprint arXiv:1612.09219},
   year = {2016},
   url = {https://arxiv.org/abs/1612.09219},
@@ -60,7 +60,7 @@
 
 @Article{autoplotly2018joss,
   author = {Tang, Yuan},
-  title = {autoplotly: An R Package for Automatic Generation of Interactive Visualizations for Statistical Results},
+  title = {autoplotly: An {R} Package for Automatic Generation of Interactive Visualizations for Statistical Results},
   journal = {Journal of Open Source Software},
   year = {2018},
   issue = {24},
@@ -71,11 +71,10 @@
 
 @Manual{autoplotlypkg,
   author = {Tang, Yuan},
-  title = {autoplotly: An R Package for Automatic Generation of Interactive Visualizations for Statistical Results},
+  title = {autoplotly: An {R} Package for Automatic Generation of Interactive Visualizations for Statistical Results},
   year = {2018},
   url = {https://CRAN.R-project.org/package=autoplotly},
 }
-
 
 @Article{lfdaoriginalpaper,
   title={Dimensionality reduction of multimodal labeled data by local fisher discriminant analysis},
