### 2018-10-20 ffc1f42
add appveyor for continuous integration
