### 2018-10-20 4f40d3f
use binary stringi for appveyor
