+++ appveyor.yml
@@ -17,6 +17,7 @@ cache:
 
 build_script:
   - travis-tool.sh install_deps
+  - travis-tool.sh install_r_binary stringi
 
 test_script:
   - travis-tool.sh run_tests
