+++ README.md
@@ -2,6 +2,7 @@
 [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/road2stat/dml?branch=master&svg=true)](https://ci.appveyor.com/project/road2stat/dml)
 [![Coverage Status](https://coveralls.io/repos/terrytangyuan/dml/badge.svg?branch=master)](https://coveralls.io/r/terrytangyuan/dml?branch=master)
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/dml)](https://cran.r-project.org/package=dml)
+[![Downloads from the RStudio CRAN mirror](https://cranlogs.r-pkg.org/badges/grand-total/dml)](https://cran.r-project.org/package=dml)
 [![License](http://img.shields.io/:license-mit-blue.svg?style=flat)](http://badges.mit-license.org)
 
 # dml (Distance Metric Learning in R)
