### 2018-10-20 e42808c
remove appveyor
