### 2018-10-20 ee98e35
Fixes #31: Remove unused docs
