+++ man/print.gmdf.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gdmf.R
+% Please edit documentation in R/gdmf.r
 \name{print.gmdf}
 \alias{print.gmdf}
 \title{Print a gmdf object}
