+++ man/GdmDiag.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gdmd.R
+% Please edit documentation in R/gdmd.r
 \name{GdmDiag}
 \alias{GdmDiag}
 \title{Global Distance Metric Learning}
@@ -35,9 +35,6 @@ For every two original data points (x1, x2) in newData (y1, y2):
 }
 \description{
 Performs Global Distance Metric Learning (GDM) on the given data, learning a diagonal matrix.
-}
-\details{
-Put GdmDiag function details here.
 }
 \note{
 Be sure to check whether the dimension of original data and constrains' format are valid for the function.
