+++ man/GdmFull.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gdmf.R
+% Please edit documentation in R/gdmf.r
 \name{GdmFull}
 \alias{GdmFull}
 \title{Global Distance Metric Learning}
@@ -33,9 +33,6 @@ For every two original data points (x1, x2) in newData (y1, y2):
 }
 \description{
 Performs Global Distance Metric Learning (GDM) on the given data, learning a full matrix.
-}
-\details{
-Put GdmFull function details here.
 }
 \note{
 Be sure to check whether the dimension of original data and constrains' format are valid for the function.
