+++ man/print.GdmDiag.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gdmd.R
+% Please edit documentation in R/gdmd.r
 \name{print.GdmDiag}
 \alias{print.GdmDiag}
 \title{Print an GdmDiag object}
