### 2018-10-20 d96a7c1
Fix RCA example
