+++ R/rca.R
@@ -21,7 +21,11 @@
 #'               Each vector represents a chunklet, the elements
 #'               in the vectors indicate where the samples locate
 #'               in \code{x}. See examples for more information.
-#'
+#' @param useD optional. When not given, RCA is done in the
+#' original dimension and \code{B} is full rank. When \code{useD} is given,
+#' RCA is preceded by constraints based LDA which reduces
+#' the dimension to \code{useD}. \code{B} in this case is of rank \code{useD}.
+#' 
 #' @return list of the RCA results:
 #' \item{B}{The RCA suggested Mahalanobis matrix.
 #'          Distances between data points x1, x2 should be
@@ -65,102 +69,128 @@
 #'
 #' @examples
 #' \dontrun{
-#' # generate synthetic Gaussian data
-#' library("MASS")
+#' library("MASS") # generate synthetic multivariate normal data
+#' set.seed(42)
+#' k <- 100L # sample size of each class
+#' n <- 3L # specify how many classes
+#' N <- k * n # total sample size
+#' x1 <- mvrnorm(k, mu = c(-16, 8), matrix(c(15, 1, 2, 10), ncol = 2))
+#' x2 <- mvrnorm(k, mu = c(0, 0), matrix(c(15, 1, 2, 10), ncol = 2))
+#' x3 <- mvrnorm(k, mu = c(16, -8), matrix(c(15, 1, 2, 10), ncol = 2))
+#' x <- as.data.frame(rbind(x1, x2, x3)) # predictors
+#' y <- gl(n, k) # response
 #'
-#' k <- 100 # sample size of each class
-#' n <- 3 # number of classes
-#' N <- k * n # total number of samples
-#'
-#' set.seed(1234)
-#' x1 <- mvrnorm(k, mu = c(-10, 6), matrix(c(10, 4, 4, 10), ncol = 2))
-#' x2 <- mvrnorm(k, mu = c(0, 0), matrix(c(10, 4, 4, 10), ncol = 2))
-#' x3 <- mvrnorm(k, mu = c(10, -6), matrix(c(10, 4, 4, 10), ncol = 2))
-#' x <- as.data.frame(rbind(x1, x2, x3))
-#' x$V3 <- gl(n, k)
-#'
-#' # The fully labeled data set with 3 classes
-#' plot(x$V1, x$V2,
-#'   bg = c("#E41A1C", "#377EB8", "#4DAF4A")[x$V3],
-#'   pch = c(rep(22, k), rep(21, k), rep(25, k))
+#' # fully labeled data set with 3 classes
+#' # need to use a line in 2D to classify
+#' plot(x[, 1L], x[, 2L],
+#'   bg = c("#E41A1C", "#377EB8", "#4DAF4A")[y],
+#'   pch = rep(c(22, 21, 25), each = k)
 #' )
-#' Sys.sleep(3)
+#' abline(a = -10, b = 1, lty = 2)
+#' abline(a = 12, b = 1, lty = 2)
 #'
-#' # The same data unlabeled; clearly the class structure is less evident
-#' plot(x$V1, x$V2)
-#' Sys.sleep(3)
+#' # generate synthetic chunklets
+#' chunks <- vector("list", 300)
+#' for (i in 1:100) chunks[[i]] <- sample(1L:100L, 10L)
+#' for (i in 101:200) chunks[[i]] <- sample(101L:200L, 10L)
+#' for (i in 201:300) chunks[[i]] <- sample(201L:300L, 10L)
 #'
-#' chunk1 <- sample(1:100, 5)
-#' chunk2 <- sample(setdiff(1:100, chunk1), 5)
[…]
