### 2018-10-20 adbdb9c
Fixes #28: Add example in README
