### 2018-10-20 499ea3f
Fix algorithm tests
