+++ tests/testthat/test_helper_functions.R
@@ -1,5 +1,0 @@
-context("helper functions")
-
-test_that("package set up successfully", {
-  sum(1, 2)
-})
