### 2018-10-20 1582266
Fixes #27: Remove references of "state-of-art"
