+++ paper/paper.md
@@ -33,9 +33,9 @@ affiliations:
 
 # Summary
 
-Distance metric is widely used in the machine learning literature. We used to choose a distance metric according to a priori (e.g. Euclidean Distance , L1 Distance, etc.) or according to the result of cross validation within small class of functions (e.g. choosing order of polynomial for a kernel). Actually, with priori knowledge of the data, we could learn a more suitable distance metric with (semi-)supervised distance metric learning techniques. `dml` [@dmlpkg] is such an R package aims to implement the state-of-the-art algorithms for (semi-)supervised distance metric learning.
+Distance metric is widely used in the machine learning literature. We used to choose a distance metric according to a priori (e.g. Euclidean Distance , L1 Distance, etc.) or according to the result of cross validation within small class of functions (e.g. choosing order of polynomial for a kernel). Actually, with priori knowledge of the data, we could learn a more suitable distance metric with (semi-)supervised distance metric learning techniques. `dml` [@dmlpkg] is such an R package aims to implement a collection of algorithms for (semi-)supervised distance metric learning.
 
-The `dml` package provides native R implementations of state-of-the-art algorithms for *Distance Metric Learning*, including both global and local methods such as *Relevant Component Analysis* [@rcapaper], *Discriminative Component Analysis* [@dcapaper], and *Local Fisher Discriminant Analysis* [@lfdaoriginalpaper]. A list of all the implemented algorithms can be found in the `dml` [package reference manual](https://cran.r-project.org/web/packages/dml/dml.pdf). These methods are widely applied in feature extraction, dimensionality reduction, clustering, information retrieval, and computer vision problems.
+The `dml` package provides native R implementations for a collection of *Distance Metric Learning* algorithms, including both global and local methods such as *Relevant Component Analysis* [@rcapaper], *Discriminative Component Analysis* [@dcapaper], and *Local Fisher Discriminant Analysis* [@lfdaoriginalpaper]. A list of all the implemented algorithms can be found in the `dml` [package reference manual](https://cran.r-project.org/web/packages/dml/dml.pdf). These methods are widely applied in feature extraction, dimensionality reduction, clustering, information retrieval, and computer vision problems.
 
 Additionally, implementations for the variants of the methods are also available in `dml` package. For example, since it was built on top of the `lfda` [@lfdapaper; @lfdapkg] package, users also have access to the family of *Local Fisher Discriminant Analysis* methods, which includes *Local Fisher Discriminant Analysis*, *Kernel Local Fisher Discriminant Analysis*, and *Semi-supervised Local Fisher Discriminant Analysis* [@semilfdapaper]. To make the results of these methods easy for users to interprete and analyze, both static and interactive visualizations for the results are available through `ggfortify` [@rjggfortify; @ggfortify] and `autoplotly` [@autoplotlypkg; @autoplotly2018joss] packages respectively.
 
