### 2018-10-20 #1036 @terrytangyuan
@joethorley Thanks! We have addressed all the issues. Please take another look to see if we missed anything else. We are waiting for another author to confirm his current affiliation but it should be ready to go once that's updated (I'll ping here).

@arfon We've also fixed the citations by putting "{}" around "R"s. I am not sure if backwards compatibility is a concern, e.g. if the previously published papers got updated and the citation style has been changed implicitly.
