### 2018-10-20 #1036 @terrytangyuan
@joethorley Okay we’ve updated the affiliations. It’s ready for another review.
