+++ paper/paper.bib
@@ -36,7 +36,7 @@
 }
 
 @Manual{dmlpkg,
-  title = {dml: Distance Metric Learning in R},
+  title = {dml: Distance Metric Learning in {R}},
   author = {Tang, Yuan and Gao, Tao and Xiao, Nan},
   year = {2015},
   note = {{R} package version 1.1.0},
