### 2018-10-20 #1036 @joethorley
@terrytangyuan and @arfon - The software and paper have passed my review. It's an elegant package - good work!
