### 2018-10-20 #1036 @terrytangyuan
Thanks for your review! I am regenerating the pdf again to populate one minor change.
