### 2018-10-23 #1036 @arfon
@terrytangyuan - a lot of the references in this submission look like they should have DOIs. Could you please update your bibtex file to include DOIs where they exist?
