### 2018-10-23 2397223
Added DOIs for some references
