+++ paper/paper.bib
@@ -12,7 +12,8 @@
   booktitle={Machine Learning for Signal Processing, 2007 IEEE Workshop on},
   pages={312--317},
   year={2007},
-  organization={IEEE}
+  organization={IEEE},
+  url={http://doi.org/10.1109/MLSP.2007.4414325},
 }
 
 @InProceedings{rcapaper,
@@ -76,14 +77,14 @@
   url = {https://CRAN.R-project.org/package=autoplotly},
 }
 
-@Article{lfdaoriginalpaper,
-  title={Dimensionality reduction of multimodal labeled data by local fisher discriminant analysis},
+@InProceedings{lfdaoriginalpaper,
+  title={Local fisher discriminant analysis for supervised dimensionality reduction},
   author={Sugiyama, Masashi},
-  journal={Journal of machine learning research},
-  volume={8},
-  number={May},
-  pages={1027--1061},
-  year={2007}
+  booktitle={Proceedings of the 23rd international conference on Machine learning},
+  pages={905--912},
+  year={2006},
+  organization={ACM},
+  url={https://doi.org/10.1145/1143844.1143958},
 }
 
 @Article{semilfdapaper,
@@ -94,5 +95,6 @@
   number={1-2},
   pages={35},
   year={2010},
-  publisher={Springer}
+  publisher={Springer},
+  url={https://doi.org/10.1007/s10994-009-5125-7},
 }
