### 2018-10-23 #1036 @terrytangyuan
@arfon Sure, I just updated the bibtex file!
