+++ paper/paper.bib
@@ -13,7 +13,7 @@
   pages={312--317},
   year={2007},
   organization={IEEE},
-  url={http://doi.org/10.1109/MLSP.2007.4414325},
+  doi={10.1109/MLSP.2007.4414325},
 }
 
 @InProceedings{rcapaper,
@@ -67,7 +67,7 @@
   issue = {24},
   volume = {3},
   pages={657},
-  url = {https://doi.org/10.21105/joss.00657}
+  doi = {10.21105/joss.00657}
 }
 
 @Manual{autoplotlypkg,
@@ -84,7 +84,7 @@
   pages={905--912},
   year={2006},
   organization={ACM},
-  url={https://doi.org/10.1145/1143844.1143958},
+  doi={10.1145/1143844.1143958},
 }
 
 @Article{semilfdapaper,
@@ -96,5 +96,5 @@
   pages={35},
   year={2010},
   publisher={Springer},
-  url={https://doi.org/10.1007/s10994-009-5125-7},
+  doi={10.1007/s10994-009-5125-7},
 }
